server code classification
